Create solution_str_slice.py
O(n + m) and O(1) space using string slicing.

Define the left and right rotate functions.

Get the mod size and the shift number, then for each direction/amount pair in the shift array, if direction is 0 (left-rotate), increment to the shift number, otherwise (if the direction is 1, i.e., right-rotate), decrement from the shift number.

Mod the shift number by the length of the string.

If the shift number is positive, left rotate, and if it is negative, right rotate. Nothing is done if it is 0.

Return the string.