Create solution.py
O(nm * min(n, m)) time and O(nm) space complexity solution using bottom-up recursion.

The first case is to simplify by swapping the dimensions if the first dimension n is greater than m.

For the special cases of terminal conditions (n = 11 and m = 13 or vice versa, whose value is 6), return 6, and similarly if the dimensions are equal, return 1.

Otherwise, create a memoization array of size (n + 1) by (m + 1).

For i and j in the range of whole numbers 1 through n (and 1 through m, respectively), if terminal conditions for index i and j are met, set it to those values and continue to the next loop iteration. Otherwise, create subproblem values of r1 and r2, as well as the minimal value variables set to inf (a MAX_VALUE that is very high equivalently in other languages). 

In an inner loop, loop x from 1 to the minimum value of the current indices (i, j) (inclusive). If j - x or i - 1 is less than 0, break out of this loop. Otherwise, set r1 to be the sum of the subproblems in the memoized array for indices (i, j-x) and (i-x, x), and r2 to be the sum of the subproblems for (i-x, j) and (x, j-x). Finally, set the minimal value to the minimum of r1, as well as the minimum of r2 and the current minimal value. 

Outside of that inner loop, set the memoization array value at (i, j) to the updated minimal value + 1.

Finally, after the n X m loop finishes, return the memoization array value at index (n, m).