Create solution_skyline.py
O(m^2 n) time and O(mn) space solution.

Use a skyline DP table and use the cache to cache DP.

Set the left pointer to 0 and the min height to a very large value.

For the index and height in the enumerated skyline tuple, if the height is less than min height, update it and the left pointer.

If the min height equals n (the max height of the whole rectangle), then return 0 (base case).

Create the new skyline as the list of the skyline. Set the result to infinity.

For the right pointer in the range from the left pointer to m (the max possible index for rectangle width), if the new skyline's value at the right pointer equals the current min height, update the tile width as the current width of right index - left index + 1. If the tile width + min height is less than or equal to n (the new square added would not go out-of-bounds), then update the heights of the new skyline from l to r inclusive with the new square's side length. Then, update the result to the minimum of the current result and the DP table value of the tuple version + 1. Otherwise, break from the loop.

Return the result from the DP function.

Run the result of the DP with the tuple of a m-sized 0-array, and return this value.